Extract component EventForm
